Refactored logic
