+++ src/Internal/Search/Searcher.php
@@ -7,19 +7,11 @@ namespace Loupe\Loupe\Internal\Search;
 use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Query\QueryBuilder;
 use Doctrine\DBAL\Result;
-use Location\Bounds;
 use Loupe\Loupe\Internal\Engine;
 use Loupe\Loupe\Internal\Filter\Ast\Ast;
-use Loupe\Loupe\Internal\Filter\Ast\Concatenator;
-use Loupe\Loupe\Internal\Filter\Ast\Filter;
-use Loupe\Loupe\Internal\Filter\Ast\GeoBoundingBox;
-use Loupe\Loupe\Internal\Filter\Ast\GeoDistance;
-use Loupe\Loupe\Internal\Filter\Ast\Group;
-use Loupe\Loupe\Internal\Filter\Ast\MultiAttributeFilter;
-use Loupe\Loupe\Internal\Filter\Ast\Node;
 use Loupe\Loupe\Internal\Filter\Parser;
 use Loupe\Loupe\Internal\Index\IndexInfo;
-use Loupe\Loupe\Internal\LoupeTypes;
+use Loupe\Loupe\Internal\Search\FilterBuilder\FilterBuilder;
 use Loupe\Loupe\Internal\Tokenizer\Token;
 use Loupe\Loupe\Internal\Tokenizer\TokenCollection;
 use Loupe\Loupe\Internal\Util;
@@ -95,26 +87,6 @@ class Searcher
         $this->geoDistanceSelectsAdded[$alias] = true;
 
         return $alias;
-    }
-
-    public function addMultiAttributeFromAndJoinToQueryBuilder(QueryBuilder $queryBuilder, string $attributeName): void
-    {
-        $queryBuilder->from(
-            IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES_DOCUMENTS,
-            $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES_DOCUMENTS)
-        )
-            ->innerJoin(
-                $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES_DOCUMENTS),
-                IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES,
-                $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES),
-                sprintf(
-                    '%s.attribute=%s AND %s.id = %s.attribute',
-                    $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES),
-                    $this->queryBuilder->createNamedParameter($attributeName),
-                    $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES),
-                    $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES_DOCUMENTS),
-                )
-            );
     }
 
     public function fetchResult(): SearchResult
@@ -379,44 +351,6 @@ class Searcher
     }
 
     /**
-     * @return array<string|float>
-     */
-    private function createGeoBoundingBoxWhereStatement(string $documentAlias, GeoBoundingBox|GeoDistance $node, Bounds $bounds): array
-    {
-        $whereStatement = [];
-
-        // Prevent nullable
-        $nullTerm = $this->queryBuilder->createNamedParameter(LoupeTypes::VALUE_NULL);
-        $whereStatement[] = $documentAlias . '.' . $node->attributeName . '_geo_lat';
-        $whereStatement[] = '!=';
-        $whereStatement[] = $nullTerm;
-        $whereStatement[] = 'AND';
-        $whereStatement[] = $documentAlias . '.' . $node->attributeName . '_geo_lng';
-        $whereStatement[] = '!=';
-        $whereStatement[] = $nullTerm;
-
-        $whereStatement[] = 'AND';
-
-        // Longitude
-        $whereStatement[] = $documentAlias . '.' . $node->attributeName . '_geo_lng';
-        $whereStatement[] = 'BETWEEN';
-        $whereStatement[] = $bounds->getWest();
-        $whereStatement[] = 'AND';
-        $whereStatement[] = $bounds->getEast();
-
-        $whereStatement[] = 'AND';
-
-        // Latitude
-        $whereStatement[] = $documentAlias . '.' . $node->attributeName . '_geo_lat';
-        $whereStatement[] = 'BETWEEN';
-        $whereStatement[] = $bounds->getSouth();
-        $whereStatement[] = 'AND';
-        $whereStatement[] = $bounds->getNorth();
-
-        return $whereStatement;
-    }
-
-    /**
      * @param array<int> $states
      */
     private function createStatesMatchWhere(array $states, string $table, string $term, int $levenshteinDistance, string $termColumnName): string
@@ -463,33 +397,6 @@ class Searcher
         );
 
         return implode(' ', $where);
-    }
-
-    /**
-     * @param array<string> $positiveMultiWheres
-     * @param array<string> $negativeMultiWheres
-     */
-    private function createSubQueryForMultiAttribute(string $attribute, array $positiveMultiWheres, array $negativeMultiWheres): string
-    {
-        $select = $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES_DOCUMENTS) . '.document';
-
-        $qb = $this->engine->getConnection()
-            ->createQueryBuilder();
-        $qb->select($select);
-
-        $this->addMultiAttributeFromAndJoinToQueryBuilder($qb, $attribute);
-
-        $qb->groupBy($select);
-
-        if ($positiveMultiWheres !== []) {
-            $qb->andHaving(sprintf('COUNT(CASE WHEN %s THEN 1 ELSE NULL END) > 0', implode(' ', $positiveMultiWheres)));
-        }
-
-        if ($negativeMultiWheres !== []) {
-            $qb->andHaving(sprintf('COUNT(CASE WHEN %s THEN 1 ELSE NULL END) = 0', implode(' ', $negativeMultiWheres)));
-        }
-
-        return $qb->getSQL();
     }
 
     private function createTermDocumentMatchesCTECondition(Token $token): ?string
@@ -616,149 +523,8 @@ class Searcher
             return;
         }
 
-        $whereStatement = [];
-
-        $this->handleFilterAstNode($this->getFilterAst()->getRoot(), $whereStatement);
-
-        $this->queryBuilder->andWhere(implode(' ', $whereStatement));
-    }
-
-    /**
-     * @param array<string|float> $whereStatement
-     */
-    private function handleFilterAstNode(Node $node, array &$whereStatement): void
-    {
-        $documentAlias = $this->engine->getIndexInfo()
-            ->getAliasForTable(IndexInfo::TABLE_NAME_DOCUMENTS);
-
-        if ($node instanceof Group) {
-            $groupWhere = [];
-            foreach ($node->getChildren() as $child) {
-                $this->handleFilterAstNode($child, $groupWhere);
-            }
-
-            if ($groupWhere !== []) {
-                $whereStatement[] = '(';
-                $whereStatement[] = implode(' ', $groupWhere);
-                $whereStatement[] = ')';
-            }
-        }
-
-        if ($node instanceof MultiAttributeFilter) {
-            $positiveMultiWheres = [];
-            $negativeMultiWheres = [];
-            foreach ($node->getChildren() as $child) {
-                // Multi filterable attributes need special handling
-                if ($child instanceof Filter && \in_array($node->attribute, $this->engine->getIndexInfo()->getMultiFilterableAttributes(), true)) {
-                    $isFloatType = LoupeTypes::isFloatType(LoupeTypes::getTypeFromValue($child->value));
-
-                    $column = $this->engine->getIndexInfo()->getAliasForTable(IndexInfo::TABLE_NAME_MULTI_ATTRIBUTES) . '.' .
-                        ($isFloatType ? 'numeric_value' : 'string_value');
-
-                    $positiveMultiWheres[] = $child->operator->buildSql(
-                        $this->engine->getConnection(),
-                        $column,
-                        $child->value
-                    );
-
-                    // If the operator is negative, we have to add a second where. This is because if the users write
-                    // $loupe->withFilter("multiattribute NOT IN ('foo', 'bar')") they want NEITHER of the attribute
-                    // values to be part of the result set. But because we store multi attributes in a separate table,
-                    // we first have to ensure the document has at least one allowed value AND then ensure the document
-                    //does NOT have ANY forbidden value
-                    if ($child->operator->isNegative()) {
-                        $negativeMultiWheres[] = $child->operator->opposite()->buildSql(
-                            $this->engine->getConnection(),
-                            $column,
-                            $child->value
-                        );
-                    }
-                } else {
-                    $this->handleFilterAstNode($child, $positiveMultiWheres);
-                }
-            }
-
-            $whereStatement[] = $documentAlias . '.id IN (';
-            $whereStatement[] = $this->createSubQueryForMultiAttribute($node->attribute, $positiveMultiWheres, $negativeMultiWheres);
-            $whereStatement[] = ')';
-        }
-
-        if ($node instanceof Filter) {
-            $operator = $node->operator;
-
-            // Not existing attributes need be handled as no match if positive and as match if negative
-            if (!\in_array($node->attribute, $this->engine->getIndexInfo()->getFilterableAttributes(), true)) {
-                $whereStatement[] = $operator->isNegative() ? '1 = 1' : '1 = 0';
-            } else {
-                $attribute = $node->attribute;
-
-                if ($attribute === $this->engine->getConfiguration()->getPrimaryKey()) {
-                    $attribute = 'user_id';
-                }
-
-                $whereStatement[] = $operator->buildSql(
-                    $this->engine->getConnection(),
-                    $documentAlias . '.' . $attribute,
-                    $node->value
-                );
-            }
-        }
-
-        if ($node instanceof GeoDistance) {
-            // Not existing attributes need be handled as no match
-            if (!\in_array($node->attributeName, $this->engine->getIndexInfo()->getFilterableAttributes(), true)) {
-                $whereStatement[] = '1 = 0';
-                return;
-            }
-
-            // Add the distance to the select query, so it's also part of the result
-            $distanceSelectAlias = $this->addGeoDistanceSelectToQueryBuilder($node->attributeName, $node->lat, $node->lng);
-
-            // Start a group
-            $whereStatement[] = '(';
-
-            // Improve performance by drawing a BBOX around our coordinates to reduce the result set considerably before
-            // the actual distance is compared. This can use indexes.
-            // We use floor() and ceil() respectively to ensure we get matches as the BearingSpherical calculation of the
-            // BBOX may not be as precise so when searching for the e.g. 3rd decimal floating point, we might exclude
-            // locations we shouldn't.
-            $bounds = $node->getBbox();
-
-            $whereStatement = [...$whereStatement, ...$this->createGeoBoundingBoxWhereStatement($documentAlias, $node, $bounds)];
-
-            // And now calculate the real distance to filter out the ones that are within the BBOX (which is a square)
-            // but not within the radius (which is a circle).
-            $whereStatement[] = 'AND';
-            $whereStatement[] = $distanceSelectAlias;
-            $whereStatement[] = '<=';
-            $whereStatement[] = $node->distance;
-
-            // End group
-            $whereStatement[] = ')';
-        }
-
-        if ($node instanceof GeoBoundingBox) {
-            // Not existing attributes need be handled as no match
-            if (!\in_array($node->attributeName, $this->engine->getIndexInfo()->getFilterableAttributes(), true)) {
-                $whereStatement[] = '1 = 0';
-                return;
-            }
-
-            // Start a group GeoDistance BBOX
-            $whereStatement[] = '(';
-
-            // Same like above for
-            $bounds = $node->getBbox();
-
-            $whereStatement = [...$whereStatement, ...$this->createGeoBoundingBoxWhereStatement($documentAlias, $node, $bounds)];
-
-            // End group
-            $whereStatement[] = ')';
-        }
-
-        if ($node instanceof Concatenator) {
-            $whereStatement[] = $node->getConcatenator();
-        }
+        $filterBuilder = new FilterBuilder($this->engine, $this, $this->queryBuilder);
+        $filterBuilder->buildForDocument();
     }
 
     /**
